requirements.txt
pymupdf
python-docx
pytesseract
Pillow
anthropic
